market_backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
SQLAlchemy==2.0.20
python-multipart==0.0.6
email-validator==2.0.0.post2
itsdangerous==2.1.2
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
